refactored some code to add hook backgroundColor class that handles background color of any object / elemets
