+++ src/components/button.tsx
@@ -1,48 +1,13 @@
 import * as React from 'react'
 import cx from 'classnames'
 import MButton, { ButtonProps } from '@material-ui/core/Button'
-import { colors, Color, Omit } from '@material-ui/core'
-import { makeStyles } from '@material-ui/styles'
+import { Omit } from '@material-ui/core'
+import { useColor } from '../hooks/useColor';
 
 export interface Props extends Omit<ButtonProps, 'color' | 'variant'> {
   confirm?: string | boolean
   color?: 'blue' | 'green' | 'red' | 'teal' | 'black' | 'inherit' | null | false
 }
-
-function makeColor(color: Color) {
-  return {
-    color: 'white',
-    background: color[400],
-    border: `1px solid ${color[400]}`,
-    transition: 'all 0.3s linear',
-    '&:hover': {
-      backgroundColor: color[300],
-      border: `1px solid ${color[300]}`,
-    },
-    '&:active': {
-      backgroundColor: color[500],
-      border: `1px solid ${color[500]}`,
-    },
-    '&:disabled': {
-      color: 'white',
-      opacity: 0.25,
-    },
-  }
-}
-
-const useStyles = makeStyles(theme => ({
-  blue: { ...makeColor(colors.blue) },
-  green: { ...makeColor(colors.green) },
-  red: { ...makeColor(colors.red) },
-  teal: { ...makeColor(colors.teal) },
-  black: {
-    color: 'black',
-    border: '1px solid black',
-  },
-  inherit: {
-    color: 'inherit',
-  },
-}))
 
 export const Button: React.SFC<Props> = ({
   onClick,
@@ -57,21 +22,15 @@ export const Button: React.SFC<Props> = ({
   const [timer, setTimer] = React.useState<NodeJS.Timeout | undefined>(
     undefined
   )
-  const classes = useStyles()
+
+  const backgroundColor = useColor(color)
 
   return (
     <MButton
       {...props}
       fullWidth={fullWidth}
       size={size}
-      className={cx(className, {
-        [classes.blue]: color === 'blue',
-        [classes.red]: color === 'red',
-        [classes.green]: color === 'green',
-        [classes.teal]: color === 'teal',
-        [classes.black]: color === 'black',
-        [classes.inherit]: color === 'inherit',
-      })}
+      className={cx(className, backgroundColor)}
       onClick={e => {
         if (onClick && confirm && !timer) {
           setTimer(
